refactor: melt and mint event payment handlers
